requirements.txt
streamlit==1.32.0
langchain==0.1.8
langchain-openai==0.0.5
python-dotenv==1.0.1
pypdf==4.0.1
docx2txt==0.8
chromadb==0.4.22
openai==1.12.0
tiktoken==0.5.2
sentence-transformers==2.2.2
langchain-community==0.0.16